requirements.txt
Django==1.4.16
MySQL-python==1.2.5
Pillow==2.7.0
South==1.0.1
argparse==1.2.1
django-extensions==1.5.0
django-mptt==0.6.1
django-registration==1.0
django-robots==1.0
django-simple-captcha==0.4.4
django-tagging==0.3.4
gunicorn==18.0
six==1.8.0
wsgiref==0.1.2
